% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{constants}
\alias{constants}
\title{Predifined Fields variables}
\format{
A set of predefined constants representing the fields used in the analysis.
}
\usage{
constants
}
\description{
These variables names should not be modified.
}
\keyword{datasets}
