% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_fishnet.R
\name{create_fishnet}
\alias{create_fishnet}
\title{Create Fishnet}
\usage{
create_fishnet(sta, square_size)
}
\arguments{
\item{sta}{an sf Simple feature collection representing the study area for analysis}

\item{square_size}{an integer value representing square side dimension, in meters}
}
\value{
an sf Simple feature collection representing a regular square fishnet
}
\description{
Creates a fishnet of regular squares over a study area
}
