% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_intersect.R
\name{create_intersect}
\alias{create_intersect}
\title{Create Intersections}
\usage{
create_intersect(nmoments, sta, field)
}
\arguments{
\item{nmoments}{a list of sf Simple feature collections representing the moments for analysis}

\item{sta}{an sf Simple feature collection representing the study area for analysis}

\item{field}{a character indicating the area field to be updated}
}
\value{
a list of sf Simple feature collections representing the moments for analysis, clipped by study area
}
\description{
Creates a list of moments for the Landscape Dynamics analysis, clipped by a study area layer.
Area field is updated for all features.
}
