% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_layouts.R
\name{create_layouts}
\alias{create_layouts}
\title{Create layouts}
\usage{
create_layouts(nmoments, in_layer)
}
\arguments{
\item{nmoments}{an integer value representing the number of moments of analysis.}

\item{in_layer}{sf Simple polygon feature collection, representing the Land Dynamics output.}
}
\value{
a list of ggplot objects, representing all Land Dynamics output layout maps.
}
\description{
Creates all possible output PNG maps.
}
