% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_moments.R
\name{create_moments}
\alias{create_moments}
\title{Create Moments}
\usage{
create_moments(momentslist, sta)
}
\arguments{
\item{momentslist}{a list of sf Simple polygon feature collections, from sf package, representing the moments for analysis, by chronological order.}

\item{sta}{a sf Simple polygon feature collection, from sf package, representing the study area for analysis.}
}
\value{
a list of sf Simple polygon feature collections representing the moments for analysis.
}
\description{
Creates a list of moments for the Landscape Dynamics analysis.
If crs of moments is not the same as the study area, a transformation is applied.
}
