% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_pairs.R
\name{create_pairs}
\alias{create_pairs}
\title{Create pairs of moments}
\usage{
create_pairs(n_moment)
}
\arguments{
\item{n_moment}{an integer value representing the number of moments for analysis.}
}
\value{
a list of vectors containing the possible pairs of moments.
}
\description{
Creates a list of all possible paring between the moments of analysis.
}
