% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_selectpatch.R
\name{create_selectpatch}
\alias{create_selectpatch}
\title{Create Select patch}
\usage{
create_selectpatch(nmoments, field, patch_value)
}
\arguments{
\item{nmoments}{a list of sf Simple feature collections representing the moments for analysis}

\item{field}{a character indicating the area field to be updated}

\item{patch_value}{an integer value representing minimum area threshold of features, in square meters}
}
\value{
a list of sf Simple feature collections representing the moments for analysis, filtered by area threshold
}
\description{
Creates a list of moments for the Landscape Dynamics analysis, discarding features with area under a threshold.
Area field is updated for all features.
}
