% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_singlepart.R
\name{create_singlepart}
\alias{create_singlepart}
\title{Create Singlepart}
\usage{
create_singlepart(nmoments, field)
}
\arguments{
\item{nmoments}{a list of sf Simple feature collections representing the moments for analysis.}

\item{field}{a character indicating the area field to be updated.}
}
\value{
a list of sf Simple feature collections representing the moments for analysis, updated to singlepart.
}
\description{
Creates a list of moments for the Landscape Dynamics analysis, updating polygons to singlepart features.
Area field is updated for all features.
}
