% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_symdif.R
\name{create_symdif}
\alias{create_symdif}
\title{Create symmetrical difference between moments}
\usage{
create_symdif(nmoments, area_field, perforation)
}
\arguments{
\item{nmoments}{a list of sf Simple feature collections representing the moments for analysis.}

\item{area_field}{a character indicating the area field to be updated.}

\item{perforation}{a logical argument indicating if perforation should be considered.}
}
\value{
a list of lists containing sf Simple feature collections representing the symmetrical differences and perforations between all moments for analysis.
}
\description{
Preform symmetrical differences and perforation between all moments of analysis.
}
