% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_areafreq.R
\name{join_areafreq}
\alias{join_areafreq}
\title{Join Area and Frequency}
\usage{
join_areafreq(nmoments, sta)
}
\arguments{
\item{nmoments}{a list of sf Simple feature collections representing the moments for analysis}

\item{sta}{an sf Simple features collection representing the study area for analysis}
}
\value{
an sf Simple features collection representing the study area for analysis, updated with area and frequency values of the different moments of analysis.
}
\description{
Updates study area's Simple feature collection by adding area and frequency fields, and populating according to the moments of analysis.
Area field is updated for all features.
}
