% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_FID_AREA.R
\name{join_FID_AREA}
\alias{join_FID_AREA}
\title{Area and Frequency Stats by feature's FID}
\usage{
join_FID_AREA(in_layer, in_df, area_field, count_field)
}
\arguments{
\item{in_layer}{an sf Simple features collection representing the study area for analysis}

\item{in_df}{a data.frame with the same amount of rows and same order as the study area's sf Simple collection, containing information on FID, sum of areas and frequency of patches.}

\item{area_field}{a character indicating the name of the field representing the sum of areas of patches in a moment of analysis}

\item{count_field}{a character indicating the name of the field representing the count of patches in a moment of analysis}
}
\value{
an sf Simple features collection representing the study area for analysis, updated with area and frequency for a moment of analysis.
}
\description{
Given a moment of analysis clipped by the study area's features, creates a data frame with FID and respective values for total area and frequency of patches.
}
