% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classLDT.R
\docType{class}
\name{Ldt-class}
\alias{Ldt-class}
\title{A Custom S4 Class for an Ldt}
\description{
This class represents all the attributes necessary to preform a Land Dynamics process.
}
\section{Slots}{

\describe{
\item{\code{analysis_squares}}{a logical value representing if the type of analysis is done by a regular grid of squares. If FALSE, the type of analysis is done by districts.}

\item{\code{nmoments}}{an integer value representing the amount of moments to consider.}

\item{\code{studyarea}}{a sf Simple polygon feature collection, from sf package, representing the study area.}

\item{\code{moments}}{a list of sf Simple polygon feature collections, from sf package, representing the moments for analysis, by chronological order.}

\item{\code{patches}}{an integer value representing minimum area threshold of features, in square meters.}

\item{\code{squares}}{an integer value representing square side dimension, in meters. Only used if type of analysis e set to squares.}

\item{\code{spatialshift}}{a logical value representing if spatial shift is to be consider.}

\item{\code{perforation}}{a logical value representing if perforation is to be consider.}

\item{\code{forecast}}{a logical value representing if forecast is to be consider.}

\item{\code{temp_fold}}{a character value representing the path for temporary folder.}

\item{\code{cluster_area}}{a numeric value representing a cluster minimum area to loop the process, to try managing memory on large data sets. Can be changed in the constructor.}

\item{\code{gap_area}}{a numeric value representing the number of features in every loop of the process, to try managing memory on large data sets. For analysis type "districts", the value is 1.}
}}

