% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_to_single.R
\name{multi_to_single}
\alias{multi_to_single}
\title{Transform multipart features into singlepart features}
\usage{
multi_to_single(in_layer)
}
\arguments{
\item{in_layer}{an sf Simple feature collection.}
}
\value{
an sf Simple feature collection.
}
\description{
Explodes all multipart features in an sf Simple feature collection.
}
