% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_perf.R
\name{select_perf}
\alias{select_perf}
\title{Select perforation features}
\usage{
select_perf(in_layer, intersect_layer)
}
\arguments{
\item{in_layer}{an sf Simple feature collection, representing symmetrical difference between a pair of moments of analysis.}

\item{intersect_layer}{an sf Simple feature collection, representing the first moment of the pair.}
}
\value{
an sf Simple feature collection representing the features where perforation was detected.
}
\description{
Selects features where perforation was detected.
}
