% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_by_FID.R
\name{stats_by_FID}
\alias{stats_by_FID}
\title{Area and Frequency Stats by feature's FID}
\usage{
stats_by_FID(in_layer)
}
\arguments{
\item{in_layer}{an sf Simple feature collection representing a moment of analysis clipped by study area's features.}
}
\value{
a data.frame with the same amount of rows and same order as the study area's sf Simple collection, containing information on FID, sum of areas and frequency of patches.
}
\description{
Given a moment of analysis clipped by the study area's features, creates a data frame with FID and respective values for total area and frequency of patches.
}
