% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_by_fid.R
\name{subset_by_fid}
\alias{subset_by_fid}
\title{Subset moments by FID}
\usage{
subset_by_fid(nmoments, fid_field, fid_select)
}
\arguments{
\item{nmoments}{a list of sf Simple feature collections representing the moments for analysis.}

\item{fid_field}{a character indicating the FID field.}

\item{fid_select}{a vector of integers, representing the FID values to subset.}
}
\value{
a list of sf Simple feature collections representing the moments for analysis, subset by FID.
}
\description{
Creates a features subset for all the moments of analysis, using FID field.
}
