% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symmetrical_difference.R
\name{symmetrical_difference}
\alias{symmetrical_difference}
\title{Symmetrical difference}
\usage{
symmetrical_difference(in_layer1, in_layer2)
}
\arguments{
\item{in_layer1}{an sf Simple feature collection.}

\item{in_layer2}{an sf Simple feature collection.}
}
\value{
an sf Simple feature collection.
}
\description{
Performs symmetrical difference between two sf Simple feature collections.
}
