% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_area.R
\name{update_area}
\alias{update_area}
\title{Update area field}
\usage{
update_area(in_layer, field)
}
\arguments{
\item{in_layer}{an sf Simple feature collection with area field.}

\item{field}{a character indicating the area field to be updated.}
}
\value{
an sf Simple feature collection, with updated area field.
}
\description{
Updates area values in sf Simple feature collection.
}
