% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_sta_perf.R
\name{update_sta_perf}
\alias{update_sta_perf}
\title{Update perforation fields}
\usage{
update_sta_perf(sta, perf_features)
}
\arguments{
\item{sta}{an sf Simple feature collection representing the study area for analysis.}

\item{perf_features}{a list of sf Simple feature collections representing the perforation between all moments for analysis.}
}
\value{
an sf Simple feature collection representing the study area for analysis, updated with perforation fields.
}
\description{
Updates perforation fields on study area's Simple feature collection.
}
