% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_sta_ToD.R
\name{update_sta_ToD}
\alias{update_sta_ToD}
\title{Update Type of Dynamics fields}
\usage{
update_sta_ToD(sta, n_moment, spatialshift, perforation, forecast)
}
\arguments{
\item{sta}{an sf Simple feature collection representing the study area for analysis.}

\item{n_moment}{an integer value representing the number of moments for analysis.}

\item{spatialshift}{a logical value representing if spatial shift is to be consider.}

\item{perforation}{a logical value indicating if perforation should be consider.}

\item{forecast}{a logical value indicating if forecast should be consider.}
}
\value{
an sf Simple feature collection representing the study area for analysis, updated with ToD fields.
}
\description{
Updates all ToD fields on study area's Simple feature collection.
}
