% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valid_moment.R
\name{valid_moment}
\alias{valid_moment}
\title{Validate moment}
\usage{
valid_moment(mm, index)
}
\arguments{
\item{mm}{a sf Simple polygon feature collection representing the moment.}

\item{index}{an integer value indicating moment index}
}
\value{
a sf Simple polygon feature collection representing the moment.
}
\description{
Check if moment's data is valid.
}
