% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valid_studyarea.R
\name{valid_studyarea}
\alias{valid_studyarea}
\title{Validate study area}
\usage{
valid_studyarea(sta)
}
\arguments{
\item{sta}{a sf Simple polygon feature collection, from sf package, representing the study area.}
}
\value{
a sf Simple polygon feature collection, from sf package, representing the study area.
}
\description{
Check if study areas's data is valid.
}
