% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_layouts.R
\name{create_layouts}
\alias{create_layouts}
\title{Create layouts}
\usage{
create_layouts(nmoments, in_layer_path)
}
\arguments{
\item{nmoments}{an integer value representing the number of moments of analysis.}

\item{in_layer_path}{a character value representing the path for the ESRI Shapefile final output.}
}
\description{
Creates all possible output PNG maps.
}
