% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_moments.R
\name{create_moments}
\alias{create_moments}
\title{Create Moments}
\usage{
create_moments(momentspaths, sta)
}
\arguments{
\item{momentspaths}{a list of ESRI Shapefiles paths representing the Moments for analysis.}

\item{sta}{an sf Simple feature collection representing the study area for analysis.}
}
\value{
a list of sf Simple feature collections representing the moments for analysis.
}
\description{
Creates a list of moments for the Landscape Dynamics analysis.
If crs of moments is not the same as the study area, a transformation is applied.
}
