% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createLDT.R
\name{createLDT}
\alias{createLDT}
\title{Create LDT object}
\usage{
createLDT(
  studyareapath,
  momentspaths,
  patches = 5000,
  squares = 10000,
  analysis_squares = T,
  spatialshift = F,
  perforation = F,
  forecast = F,
  output = NULL
)
}
\arguments{
\item{studyareapath}{a character value representing the path for the ESRI Shapefile that defines the study area.}

\item{momentspaths}{a vector of character values representing the paths for the ESRI Shapefiles that defines the moments for analysis, by chronological order.}

\item{patches}{an integer value representing minimum area threshold of features, in square meters. Default to 5000.}

\item{squares}{an integer value representing square side dimension, in meters. Only used if type of analysis if set to squares. Default to 10000.}

\item{analysis_squares}{a logical value representing the type of analysis. If TRUE, the analysis is done by a regular grid of squares. If FALSE, the analysis is done by districts. Default to TRUE.}

\item{spatialshift}{a logical value representing if spatial shift is to be consider. Default to FALSE.}

\item{perforation}{a logical value representing if perforation is to be consider. Default to FALSE.}

\item{forecast}{a logical value representing if forecast is to be consider. Default to FALSE.}

\item{output}{a character value representing the path for the ESRI Shapefile output. Optional.}
}
\value{
an LDT object.
}
\description{
Constructor for LDT objects.
}
\examples{
\donttest{

# create a 2 moments 30000 meters square size Land dynamics object, considering patches over 1000 square meters, spatial shift, perforation and forecast
objLDT <- createLDT("/path_to_wd/studyarea.shp", c("/path_to_wd/moment1.shp", "/path_to_wd/moment2.shp"), patches=1000, squares=30000, analysis_squares=T, spatialshift=T, perforation=T, 
forecast=T, output="/path_to_output/outLDT.shp")
}
}
