% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createLDT.R
\name{createLDT}
\alias{createLDT}
\title{Create LDT object}
\usage{
createLDT(
  studyarea,
  moments,
  patches = 5000,
  squares = 10000,
  analysis_squares = T,
  spatialshift = F,
  perforation = F,
  forecast = F
)
}
\arguments{
\item{studyarea}{a sf Simple polygon feature collection, from sf package, representing the study area.}

\item{moments}{a list of sf Simple polygon feature collections, from sf package, representing the moments for analysis, by chronological order.}

\item{patches}{an integer value representing minimum area threshold of features, in square meters. Default to 5000.}

\item{squares}{an integer value representing square side dimension, in meters. Only used if type of analysis if set to squares. Default to 10000.}

\item{analysis_squares}{a logical value representing the type of analysis. If \code{TRUE}, the analysis is done by a regular grid of squares. If \code{FALSE}, the analysis is done by districts. Default to \code{TRUE}.}

\item{spatialshift}{a logical value representing if spatial shift is to be consider. Default to \code{FALSE}.}

\item{perforation}{a logical value representing if perforation is to be consider. Default to \code{FALSE}.}

\item{forecast}{a logical value representing if forecast is to be consider. Default to \code{FALSE}.}
}
\value{
an LDT object.
}
\description{
Constructor for LDT objects.
}
\examples{
\donttest{
library(sf)

path_gpkg <- system.file(
  "extdata",
  "example_ldt.gpkg",
  package = "ldtr"
)

std_area <- st_read(
  path_gpkg,
  layer = "study_area",
  quiet = TRUE
)

list_moments <- list(
  st_read(path_gpkg, layer = "moment1", quiet = TRUE),
  st_read(path_gpkg, layer = "moment2", quiet = TRUE)
)


# Create a 2 moments 10000 meters square size Land dynamics object, 
considering patches over 1000 square meters, 
spatial shift, perforation and forecast

objLDT <- createLDT(std_area, 
                    list_moments, 
                    patches=1000,
                    spatialshift=T, 
                    perforation=T, 
                    forecast=T)
}
}
