% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createLDT.R
\name{createLDT}
\alias{createLDT}
\title{Create LDT object}
\usage{
createLDT(
  analysis_squares,
  nmoments,
  studyareapath,
  momentspaths,
  patches,
  squares,
  spatialshift,
  perforation,
  forecast,
  output
)
}
\value{
an LDT object.
}
\description{
Constructor for LDT objects.
}
\section{Slots}{

\describe{
\item{\code{analysis_squares}}{a logical value representing if the type of analysis is done by a regular grid of squares. If false, the type of analysis is done by districts.}

\item{\code{nmoments}}{an integer value representing the amount of moments to consider.}

\item{\code{studyareapath}}{a character value representing the path for the ESRI Shapefile that defines the study area.}

\item{\code{momentspaths}}{a vector of character values representing the paths for the ESRI Shapefiles that defines the moments for analysis, by chronological order.}

\item{\code{patches}}{an integer value representing minimum area threshold of features, in square meters.}

\item{\code{squares}}{an integer value representing square side dimension, in meters. Only used if type of analysis if set to squares.}

\item{\code{spatialshift}}{a logical value representing if spatial shift is to be consider.}

\item{\code{perforation}}{a logical value representing if perforation is to be consider.}

\item{\code{forecast}}{a logical value representing if forecast is to be consider.}

\item{\code{output}}{a character value representing the path for the ESRI Shapefile output.}
}}

\examples{

# set working directory
setwd("/path_to_wd/")

# create a 2 moments 30000 meters square size Land dynamics object, considering patches over 1000 square meters, spatial shift, perforation and forecast
objLDT <- createLDT(T, 2, "studyarea.shp", c("moment1.shp", "moment2.shp"), 1000, 30000, T, T, T, "outLDT.shp")

}
