% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valid_moment.R
\name{valid_moment}
\alias{valid_moment}
\title{Validate moment}
\usage{
valid_moment(mm_path, index)
}
\arguments{
\item{mm_path}{an ESRI Shapefile path representing the moment.}
}
\value{
an sf Simple feature collection representing the moment.
}
\description{
Read moment data and check if it's valid.
}
