% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valid_studyarea.R
\name{valid_studyarea}
\alias{valid_studyarea}
\title{Validate study area}
\usage{
valid_studyarea(sta_path)
}
\arguments{
\item{sta_path}{an ESRI Shapefile path representing the study area.}
}
\value{
an sf Simple feature collection representing the study area.
}
\description{
Read study area data and check if it's valid.
}
