% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeLDT.R
\name{writeLDT}
\alias{writeLDT}
\title{Write LDT output}
\usage{
writeLDT(
  objLDT,
  saveoutput = F,
  savemaps = F,
  layername = NULL,
  outfolder = NULL
)
}
\arguments{
\item{objLDT}{an object of class \code{LDT}.}

\item{saveoutput}{a logical value representing if output layer should be exported. Default to \code{FALSE}. If \code{TRUE}}

\item{savemaps}{a logical value representing if output maps should be exported. Default to \code{FALSE}.}

\item{outfolder}{a character value representing the path for output folder. Used only when \code{savemaps = TRUE}}
}
\value{
a list with 2 elements: a sf Simple polygon feature collection representing the Land Dynamics output, and a list of ggplot objects, representing all Land Dynamics output layout maps.
}
\description{
Performs the Landscape Dynamics process and returns a sf Simple polygon feature collection and a list of ggplot maps.
}
\examples{
\donttest{
library(sf)

path_gpkg <- system.file(
  "extdata",
  "example_ldt.gpkg",
  package = "ldtr"
)

std_area <- st_read(
  path_gpkg,
  layer = "study_area",
  quiet = TRUE
)

list_moments <- list(
  st_read(path_gpkg, layer = "moment1", quiet = TRUE),
  st_read(path_gpkg, layer = "moment2", quiet = TRUE)
)


# Create a 2 moments 10000 meters square size Land dynamics object, 
considering patches over 1000 square meters, 
spatial shift, perforation and forecast

objLDT <- createLDT(std_area, 
                    list_moments, 
                    patches=1000,
                    spatialshift=T, 
                    perforation=T, 
                    forecast=T)

# run LDT
writeLDT(objLDT)
}
}
