% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeLDT.R
\name{writeLDT}
\alias{writeLDT}
\title{Write LDT output}
\usage{
writeLDT(objLDT)
}
\value{
an sf Simple feature collection, representing the Land Dynamics output.
}
\description{
Preforms the Land Dynamics process and exports an ESRI Shapefile and png maps as outputs.
}
\section{Slots}{

\describe{
\item{\code{objLDT}}{an object of class LDT.}
}}

