% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeLDT.R
\name{writeLDT}
\alias{writeLDT}
\title{Write LDT output}
\usage{
writeLDT(objLDT)
}
\value{
an sf Simple feature collection, representing the Land Dynamics output.
}
\description{
Performs the Landscape Dynamics process and exports an ESRI Shapefile and png maps as outputs.
}
\section{Slots}{

\describe{
\item{\code{objLDT}}{an object of class LDT.}
}}

\examples{

# set working directory
setwd("/path_to_wd/")

# create a 2 moments 30000 meters square size Land dynamics object, considering patches over 1000 square meters, spatial shift, perforation and forecast
objLDT <- createLDT(T, 2, "studyarea.shp", c("moment1.shp", "moment2.shp"), 1000, 30000, T, T, T, "outLDT.shp")

# run LDT
writeLDT(objLDT)

}
